requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
websockets==12.0
aiosqlite==0.19.0
python-dotenv==1.0.0
apscheduler==3.10.4
pydantic==2.5.3
pydantic-settings==2.1.0
jinja2==3.1.3
python-multipart==0.0.6
aiosqlite
apscheduler
fastapi
httpx
jinja2
pydantic
pydantic-settings
python-dotenv
python-multipart
uvicorn[standard]
websockets
